The first versions of a service base spellchecker.
spelld:
  This is the spell checker server which follows more or less the Ikaki
protocol.  This will likely change as there are some shortcommings in
that protocol.  The spell server will listen on a specified port and
answer spelling queries.

spellchecker:
  This class for now only contains wrappers for Microsoft Spell checking software
but the idea is to place of classes here to allow any number of backend
spell checkers to be used.  This class is used by spelld but it could
also be used by almot any other program.

TODO:
  * Make the backend selectable
  * Look at a proper Python test client
  * Reevaluate the protocol to manage things like: spelling capabilities,
    password, just checking vs getting suggestions.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/trunk@1558 54714841-351b-0410-a198-e36a94b762f5