Recovering from SF CVS crash: Origianl 1.35 commit: The problem being corrected occurs when start and end delimeters are not the same lenght eg in HTML <!-- (4) --> (3).  In this corner we have a line '[space]-->' and we are already in the string.  So lastpos != pos = lenstart is 0 != 4 - 4 ie we get confused.
I think this fix is correct in that it ignores the case when our last delimeter was at pisition 0.  Because
we only do this when we are instring I don't think it will mess with other cases.

Original 1.36 commit:
rather fix this by checking whether lastpos really is a startdelim...


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/trunk@2480 54714841-351b-0410-a198-e36a94b762f5