This is a big commit. The changes are (mostly) interdependent, but can be described individually.
1. Split the cursor functionality in virtaal.support.set_enumerator into its own
class. The motivation is that it makes the code easier to understand; as a nice
side effect, numerous cursors into a single UnionSetEnumerator can now be held.

2. Added a mode box to the status bar. This required a minor change in virtaal.glade.
It also necessitated the addition of a widget called ModeBox. This widget is
responsible for showing which mode is currently selected and for generating 
appropriate GObject signals when the user selects a signal.
2.1. Added mode switching code to virtaal.document. This code works along with the
code described above.

3. Various convenience methods added to virtaal.unit_grid; these make it less
confusing to switch between the treeview paths and translation store indices.
3.1. As a consequence, the keyboard and mouse navigation code was improved. These
improvements are also due to the improvements in the mode selection code.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk@7294 54714841-351b-0410-a198-e36a94b762f5