Got rid of the horrible height calculation code.
Before, VirTaal computed the heights of the various widgets in the
editing area so that it could set the editing area to the correct
size.

It turns out that one only needs to compute the and set the
heights of the text boxes, since they have no way of knowing that
one might want to display all the text in them.

Then one simply calls size_request() on the top-most editing widget
and presto! the correct height for the editing widget is computed.

To set the heights of the widgets, we use the generic method
system. We only bother setting heights on text boxes and comment
labels.

Also, please note that the code for creating editor widgets has
been factored out into get_editor.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk@7459 54714841-351b-0410-a198-e36a94b762f5