StatsDb._getstoredfileid could return None. This was a problem in methods which used _getstoredfileid, since some expected the return value not to be None.
It turns out that there is no use for a None return value. So if there is no
stored fileid, then we simply generate it with cachestore. Since we're not just
returning stored fileids, _getstoredfileid was renamed to _getfileid. Naturally,
those methods that handled the case where None was returned could have that code
removed.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk@7508 54714841-351b-0410-a198-e36a94b762f5