Implemented work-flow code in gui.EditArea. It works, although it can still easily be broken.
When a word (surface form) is selected in the word list, the edit area is
notified and loads the associated root and part of speech if applicable (the
surface will usually not have any such association). The focus is given to
the root selection combo box. After a root is selected or a new one entered,
the focus moves to the part of speech combo box. This will usually not change
and will usually not be allowed to change. If a new root was entered, the
"Add Root" button is shown and must be clicked before the loaded surface form
can be classified. Once "Add Root" is clicked the language database is updated
with the changes and the "Add Root" button is replaced by the "OK"
(classification) button. If an existing root was selected, but with a different
part of speech, both the "Add Root" and "Modify Root" buttons will be shown.
In this case "Add Root" will create a duplicate root with a different part of
speech. "Modify Root" will update the existing root's associated part of speech.
The "OK" button, like "Reject" and "Ignore", updates the current surface form's
status and refreshes the word list (which causes the edit area to load the next
word).

The current implementation relies heavily on the ComboBoxEntryies' Entries'
"activate" event (comboboxentry.child.connect("activate", ...)) for updating
the GUI and can be broken by using tab or the mouse. At least there's now
something to work with.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk@7561 54714841-351b-0410-a198-e36a94b762f5