Due to the original design of Pootle, pootlefile was inherited from the po file class. When XLIFF support was added to Pootle, pootlefile's base class now had to be either a pofile or an xliff file.
This posed some problems. It was originally solved by storing the store object inside
a reference called __innerobj__ and pootlefile would delegate requests to this object.
Profiling showed that some of the delegation machinary (especially __setattr__
requests) on Wrapper took a lot of time.

Instead of maintaining __innerobj__, we now set the base class of pootlefile at runtime
according to store (po or xliff) we are wrapping. This means that delegation happens
automatically, since pootlefile is now just a subclass of the store it contains.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/branches/Pootle-diet-and-trunk-merge@7611 54714841-351b-0410-a198-e36a94b762f5