Before the statsdb changes were incorporated, stats for a filename (computed in TranslationProject.combinestats) were stored as arrays of unit indices. For example, there would be an array called 'fuzzy', which stored the indices of fuzzy units.
With the change to the statsdb code, the code in combinestats (which
remained unchanged), called getpototlas, which now only returned
integers instead of arrays.

This broke the code which assumed to find arrays. In addition, only
a few totals were returned in this fashion and not everything that
Pootle needs to know.

combinestats returns a dictionary indexed by property names. This
patch augments this dictionary by adding the key 'units', which
is itself a dictionary containing arrays of unit indices as its
values. The dictionary keys are the property names, such as 'fuzzy'.
This gives access to per-unit properties in a file, as before.

In the previous code, assignments were also stored in the dictionary
returned by combinestats, by prepending every assign with the string
"assign-". The assignments have also been moved to their own
dictionary (as the units above). There is now an entry "assign" in
the dictionary returned by combinestats which is a dictionary of
assigns.

Code was patched up in indexpage.py to accomodate these changes.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/branches/Pootle-diet-and-trunk-merge@7613 54714841-351b-0410-a198-e36a94b762f5