Two things:
1. If we set the terminology directory config option before the
config file is read, it will be overwritten. This was changed in
run_virtaal.py so that the terminology directory is only set
after the config file has been read.

2. The terminology matcher already has support for building a
matcher from multiple files. So we don't need to build such a
file; we only need to pass it a list of stores and it will do
the rest.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk@7922 54714841-351b-0410-a198-e36a94b762f5