Added a customized entry-completion to the "root word" gtk.Entry.
The customization involves having a delay (EditArea.COMPLETION_POPUP_DELAY) to
keep from having to look for matches as soon as a character is typed. A limit
on the number of completion candidates are also imposed
(EditArea.MAX_COMPLETION_LENGTH) to make sure that we don't keep the processor
busy longer than what is useful.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk@8001 54714841-351b-0410-a198-e36a94b762f5