Change the indexing system so that we have to ask for project.indexer if we want to do an indexing query.
project.indexer computes the index of a project lazily. Thus, if it
is invoked, it first checks to see whether a project's index is up to
date (by calling project.initindex).

project.indexer returns None if there is no indexer for a project. This
can happen if indexing is not available (that is, if indexer.HAVE_INDEX
is false) or if the index for a project is corrupt. In this way we
guard against corrupt indices making Pootle unusable for some projects.

The --refreshstats functionality was also updated to force index 
initialization, which is required due to the laziness.



git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk@8198 54714841-351b-0410-a198-e36a94b762f5