This is an initial replacement of the Python based PO parser.
This parser causes a few unit tests to fail which don't fail with
the existing Python PO parser. This is mostly because it does not
treat whitespace lines in exactly the same way as the previous
parser. THIS SHOULD BE FIXED.

Currently, the parser lacks comments. These are to be added in the
next commit.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/branches/translate@8317 54714841-351b-0410-a198-e36a94b762f5