This is some fairly dense code to extract XML from an arbitrary document, given tables of 1. XML namespaces which should be converted, and 2. XML namespaces which should appear as placables within other    translatable hunks.
The code works by searching through the DOM tree (the function apply),
until it hits a translatable tag. Then it calls process_translatable_tag.

process_translatable_tag sees whether there are any sub-translatable
tags in the current translatable tag. An example of this is footnotes
in OpenOffice.org documents. The XML code for a footnote appears within
the paragraph tag with which the footnote is associated. Or in XML:

<text:p text:style-name="Standard">First. This should
  <text:note text:id="ftn0" text:note-class="footnote">
    <text:note-citation>1</text:note-citation>
    <text:note-body>
      <text:p text:style-name="Footnote">Footnote 1</text:p>
    </text:note-body>
  </text:note>not be segmented. Even with etc. and so.
</text:p>

We need to treat tags like <text:note> as placables, which means that
the above should be presented to the translator as something like:

  First. This should&footnote_1; not be segmented. Even with etc. and so.

Note that the entire XML block related to the footnote is represented
by:

  &footnote_1;

Thus, process_translatable_tag is responsible for finding any placables
in the current translatable_tag. If there are any placables, it should
create placable tags for them (such as &footnote_1;) and construct a
translatable string containing these placable tags. Then it proceeds to
deal with the children (that is, the placables) by invoking apply on them.

Note that the current implementation uses the Record type, which provides
immutable records (for stateless programming). Because it is immutable,
every modification creates a new record.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/branches/translate@8320 54714841-351b-0410-a198-e36a94b762f5