Maintain a stack of placeable names in the parse state.
Recall that a placeable might be a whole nested XML structure. We
might want to use a tag somewhere in the middle of this structure
to name the placeable.

Thus, when we hit a tag, we check whether it appears in
the placeable_table. If so, we push a name onto the placeable stack.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/branches/translate@8340 54714841-351b-0410-a198-e36a94b762f5