Resurrected XPathBreadcrumb and put it to work in the parse state.
It replaces the explicit xpath stack. Instead we call start_tag and
end_tag. This class takes care of keeping track of the number o
occurrences of a given tag (so that it can give an index to a tag).

We now also store the full XPath of a translatable as a string,
which we get from the breadcrumb.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/branches/translate@8343 54714841-351b-0410-a198-e36a94b762f5