Modified storage/odf_shared.py to attempt first to import itools and to use its ODF information. Failing that, it falls back to a copy of the itools information in storage/odf_shared.py (which may be out of date).
The important change is that we initially listed the tags in which we
were interested, whereas itools lists tags that should be ignored.

Due to integration of the code with itools, the specification mechanism
has also been simplified. We only have a table of tags we reject and a
table of inline placeables.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk@8750 54714841-351b-0410-a198-e36a94b762f5