When loading a file, we ask the statistics db module to give us an array of all the units in a file which are editable.
Once this is loaded, it remains loaded, until we reload a file,
but not when we saved a file.

If the user edited a PO file without a header, Pootle would add
a header and save the file - this would add a new unit to the
top of the file. But since we didn't update the list of
editable units when saving, we had an incorrect view of the file.
We fix this by forcing the list of editable units to be reloaded
when saving.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/branches/Pootle-toolkit-1.2@9049 54714841-351b-0410-a198-e36a94b762f5