Remove the fake "refreshing" of the StoreCursor after assigning to the cursor's indices list. It didn't work to begin with, but is now handled by the cursor's indices property anyway.
Note the assumption documented in SearchMode.selected(): It is assumed that,
whenever a file is loaded, the mode's selected() is called (to give the mode
a chance to do its thing on the new file). This is why
ModeController._select_mode() no longer returns immediately if the parameter
mode is the same as the current mode.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/branches/virtaal@9123 54714841-351b-0410-a198-e36a94b762f5