Here we provide Django ORM implementations of the SQLAlchemy models.
In addition to the SQLAlchemy models, there is a model called
PootleProfile. This is used to tie application-specific user
data to the generic Django User model (which contains basic
information such as the username, password, e-mail address,
superuser status).

Note the use of managers (LanguageManager, ProjectManager,
SubmissionManager, SuggestionManager) to implement queries
which cannot be expressed in terms of Django's ORM primitives.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk@9207 54714841-351b-0410-a198-e36a94b762f5