Backport:
r8916
Adapt dialect description to correct problems with csv pre Python 2.5.  We retain
2.5 behaviour and only adjust on older version, so that we can protect most
users from potential brokenness.

r8914
Default to using Latin1 instead of UTF-16.  We had the problem that in po2wordfast we always produce UTF-16 files.
This was because the wordfast files where initiated to be utf-16 (we only did the right thing if we parsed an 
existing wordfast file).


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/branches/Pootle-toolkit-1.2@9219 54714841-351b-0410-a198-e36a94b762f5