Convert OOo help escaping, resolves bug 694
Previously we escaped all of OOo help tags except tags that we
identified as unescaped.  This caused problems where translators
introduced new 'tags' in their transations.

We now don't escape anything unless it is on our list of tags
that should be escaped.  A translator who uses one of these tags,
which could be a valid word in their language, will still suffer.
But that is unlikely to happen and can be worked around.


git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk@9950 54714841-351b-0410-a198-e36a94b762f5