This patch removes the need for the POTree class (the file potree.py will be removed in the near future).
Most of the interesting code in potree.py has been moved to
pootle_app/project_tree.py.

This patch marks a shift in the Pootle code. The preference is now
to use Project and Language objects directly, instead of using
project and language codes. In Pootle 1.2, this made sense, since
Pootle maintained dictionaries of projects and languages, which were
indexed by project and language codes respectively. Since we're
reading from a database now, this scheme is unnecessarily slow.

We should try to read Language and Project objects once and then
use them throughout the code. Obviously, Pootle has some way to
go to reach this kind of efficiency, but we'll get there.

**

This patch has also started sticking to the convention of calling
an specific instance of a project a "translation project". Thus,
"pootle" is a "project", but the French instance of "pootle" is
a "translation project". Although this terminilogy is not ideal,
it matches historical usage and we should stick to it to avoid
confusion.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk@10169 54714841-351b-0410-a198-e36a94b762f5