Replaced the permissions management screen with a Django forms-based solution.
This will only display permissions which have been defined for the
directory of the current TranslationProject.

We should modify this so that it can be become the permissions page
to edit permissions for any directory.

We also need a way to display the effective permissions for a given
user, so that the admin will know who has what permissions.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk@10333 54714841-351b-0410-a198-e36a94b762f5