Make it possible to work atomically on multiple stores at the same time.
This introduces a possible deadlock scenario. We should be very careful
about locking here.

If two different processes each tries to work on a set of stores
and they share stores, then if we do locking, we might get them to
deadlock as they wait for each other to release their locks.

The way to deal with this is to have a global lock that must be gained
before individual files may be locked.

This problem will disappear (or at least be deferred to the file system's
file locking machinery) when Pootle becomes stateless (which is
sooner than you think).

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk@10425 54714841-351b-0410-a198-e36a94b762f5