Load the first translate page directly with JavaScript.
Lots of code that relied on the old translate page has been commented,
so this may need a cleanup/reorganisation.

This now improves load time, avoids unnecessary processing and kind of
requires JavaScript. We may need an indicator on the first page load to
notice that something is going on.

git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/src/trunk@16583 54714841-351b-0410-a198-e36a94b762f5